Add first draft of the `deploy host` command
This is based off of an earlier version of a golem script that did most of the heavy lifting for setting up a new lxc host.

The roadmap for biphrost includes auto-scaling lxc hosts, so a command like this one will be needed, but a lot more work is yet to be done to get there.

TODO:
* Add API integrations for VPS providers
* Use Incus
* Swap out the LetsEncrypt script for the new `ssl update` command
* Fix a potential hole in lxc networking configuration